requirements.txt
Django==5.2.1
psycopg2-binary==2.9.10
pillow==11.2.1
djangorestframework==3.15.2
python-decouple==3.8
Pillow==11.2.1
django-filter==24.3
django-cors-headers==4.6.0
redis==5.0.8
celery==5.4.0
whitenoise==6.8.2
gunicorn==23.0.0
google-generativeai==0.8.3
google-genai
requests==2.32.3
markdown==3.8
Pygments==2.19.1